mcp/mcp/gemini-yacht-mcp/requirements.txt
# MCP Server Dependencies
mcp>=1.2.0

# Google Gemini API
google-generativeai>=0.8.0

# Validation & Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Async HTTP
aiohttp>=3.9.0

# Utilities
typing-extensions>=4.8.0
